requirements.txt
anyascii==0.3.2
arrow==1.3.0
asgiref==3.8.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cr==1.10
cryptography==45.0.4
defusedxml==0.7.1
Django==5.1.11
django-allauth==65.9.0
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.1
django-taggit==6.1.0
django-tailwind==4.0.1
django-tasks==0.6.1
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
draftjs_exporter==5.1.0
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
fsspec==2025.5.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
laces==0.1.2
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
paramiko==3.5.1
pillow==11.2.1
pillow_heif==0.21.0
psycopg==3.2.9
pyclipper==1.3.0.post6
pycparser==2.22
Pygments==2.19.2
pyjson==1.4.1
PyMuPDF==1.26.1
PyNaCl==1.5.0
PyPDF2==3.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==13.9.4
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
telepath==0.3.1
text-unidecode==1.3
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wagtail==6.4.2
Willow==1.10.0
